+++ lib/api_trakt/indexerapiinterface.py
@@ -441,14 +441,7 @@ class TraktIndexer(TVInfoBase):
     def _get_show_lists(self, url, account=None):
         # type: (str, Any) -> List[TVInfoShow]
         result = []
-        if account:
-            from sickgear import TRAKT_ACCOUNTS
-            if account in TRAKT_ACCOUNTS and TRAKT_ACCOUNTS[account].active:
-                kw = {'send_oauth': account}
-            else:
-                raise TraktAuthException('Account missing or disabled')
-        else:
-            kw = {}
+        kw = {}
         resp = TraktAPI().trakt_request(url, **kw)
         if resp:
             for _show in resp:
@@ -463,136 +456,37 @@ class TraktIndexer(TVInfoBase):
         :param result_count: how many results are suppose to be returned
         """
         use_period = ('weekly', period)[period in ('daily', 'weekly', 'monthly', 'yearly', 'all')]
-        return self._get_show_lists('shows/played/%s?extended=full&page=%d&limit=%d' % (use_period, 1, result_count))
+        return self._get_show_lists(f'shows/played/{use_period}?extended=full&page={1:d}&limit={result_count:d}')
 
     def get_most_watched(self, result_count=100, period='weekly', **kwargs):
         # type: (...) -> List[TVInfoShow]
         """
         get most watched shows
         :param period: possible values: 'daily', 'weekly', 'monthly', 'yearly', 'all'
-        :param result_count: how many results are suppose to be returned
+        :param result_count: how many results are supposed to be returned
         """
         use_period = ('weekly', period)[period in ('daily', 'weekly', 'monthly', 'yearly', 'all')]
-        return self._get_show_lists('shows/watched/%s?extended=full&page=%d&limit=%d' % (use_period, 1, result_count))
+        return self._get_show_lists(f'shows/watched/{use_period}?extended=full&page={1:d}&limit={result_count:d}')
 
     def get_most_collected(self, result_count=100, period='weekly', **kwargs):
         # type: (...) -> List[TVInfoShow]
         """
         get most collected shows
         :param period: possible values: 'daily', 'weekly', 'monthly', 'yearly', 'all'
-        :param result_count: how many results are suppose to be returned
+        :param result_count: how many results are supposed to be returned
         """
         use_period = ('weekly', period)[period in ('daily', 'weekly', 'monthly', 'yearly', 'all')]
-        return self._get_show_lists('shows/collected/%s?extended=full&page=%d&limit=%d' % (use_period, 1, result_count))
+        return self._get_show_lists(f'shows/collected/{use_period}?extended=full&page={1:d}&limit={result_count:d}')
 
     def get_recommended(self, result_count=100, period='weekly', **kwargs):
         # type: (...) -> List[TVInfoShow]
         """
         get most recommended shows
         :param period: possible values: 'daily', 'weekly', 'monthly', 'yearly', 'all'
-        :param result_count: how many results are suppose to be returned
+        :param result_count: how many results are supposed to be returned
         """
         use_period = ('weekly', period)[period in ('daily', 'weekly', 'monthly', 'yearly', 'all')]
-        return self._get_show_lists('shows/recommended/%s?extended=full&page=%d&limit=%d' % (use_period, 1, result_count))
-
-    def get_recommended_for_account(self, account, result_count=100, ignore_collected=False, ignore_watchlisted=False,
-                                    **kwargs):
-        # type: (...) -> List[TVInfoShow]
-        """
-        get most recommended shows for account
-        :param account: account to get recommendations for
-        :param result_count: how many results are suppose to be returned
-        :param ignore_collected: exclude colleded shows
-        :param ignore_watchlisted: exclude watchlisted shows
-        """
-        from sickgear import TRAKT_ACCOUNTS
-        if not account or account not in TRAKT_ACCOUNTS or not TRAKT_ACCOUNTS[account].active:
-            raise TraktAuthException('Account missing or disabled')
-        extra_param = []
-        if ignore_collected:
-            extra_param.append('ignore_collected=true')
-        if ignore_watchlisted:
-            extra_param.append('ignore_watchlisted=true')
-        return self._get_show_lists('recommendations/shows?extended=full&page=%d&limit=%d%s' %
-                                    (1, result_count, ('', '&%s' % '&'.join(extra_param))[0 < len(extra_param)]),
-                                    account=account)
-
-    def hide_recommended_for_account(self, account, show_ids, **kwargs):
-        # type: (integer_types, List[integer_types], Any) -> List[integer_types]
-        """
-        hide recommended show for account
-        :param account: account to get recommendations for
-        :param show_ids: list of show_ids to no longer recommend for account
-        :return: list of added ids
-        """
-        from sickgear import TRAKT_ACCOUNTS
-        if not account or account not in TRAKT_ACCOUNTS or not TRAKT_ACCOUNTS[account].active:
-            raise TraktAuthException('Account missing or disabled')
-        if not isinstance(show_ids, list) or not show_ids or any(not isinstance(_i, int) for _i in show_ids):
-            raise TraktException('list of show_ids (trakt id) required')
-        resp = TraktAPI().trakt_request('users/hidden/recommendations', send_oauth=account,
-                                        data={'shows': [{'ids': {'trakt': _i}} for _i in show_ids]})
-        if resp and isinstance(resp, dict) and 'added' in resp and 'shows' in resp['added']:
-            if len(show_ids) == resp['added']['shows']:
-                return show_ids
-            if 'not_found' in resp and 'shows' in resp['not_found']:
-                not_found = [_i['ids']['trakt'] for _i in resp['not_found']['shows']]
-            else:
-                not_found = []
-            return [_i for _i in show_ids if _i not in not_found]
-        return []
-
-    def unhide_recommended_for_account(self, account, show_ids, **kwargs):
-        # type: (integer_types, List[integer_types], Any) -> List[integer_types]
-        """
-        unhide recommended show for account
-        :param account: account to get recommendations for
-        :param show_ids: list of show_ids to be included in possible recommend for account
-        :return: list of removed ids
-        """
-        from sickgear import TRAKT_ACCOUNTS
-        if not account or account not in TRAKT_ACCOUNTS or not TRAKT_ACCOUNTS[account].active:
-            raise TraktAuthException('Account missing or disabled')
-        if not isinstance(show_ids, list) or not show_ids or any(not isinstance(_i, int) for _i in show_ids):
-            raise TraktException('list of show_ids (trakt id) required')
-        resp = TraktAPI().trakt_request('users/hidden/recommendations/remove', send_oauth=account,
-                                        data={'shows': [{'ids': {'trakt': _i}} for _i in show_ids]})
-        if resp and isinstance(resp, dict) and 'deleted' in resp and 'shows' in resp['deleted']:
-            if len(show_ids) == resp['deleted']['shows']:
-                return show_ids
-            if 'not_found' in resp and 'shows' in resp['not_found']:
-                not_found = [_i['ids']['trakt'] for _i in resp['not_found']['shows']]
-            else:
-                not_found = []
-            return [_i for _i in show_ids if _i not in not_found]
-        return []
-
-    def list_hidden_recommended_for_account(self, account, **kwargs):
-        # type: (integer_types, Any) -> List[TVInfoShow]
-        """
-        list hidden recommended show for account
-        :param account: account to get recommendations for
-        :return: list of hidden shows
-        """
-        from sickgear import TRAKT_ACCOUNTS
-        if not account or account not in TRAKT_ACCOUNTS or not TRAKT_ACCOUNTS[account].active:
-            raise TraktAuthException('Account missing or disabled')
-        return self._get_show_lists('users/hidden/recommendations?type=show', account=account)
-
-    def get_watchlisted_for_account(self, account, result_count=100, sort='rank', **kwargs):
-        # type: (...) -> List[TVInfoShow]
-        """
-        get watchlisted shows for the account
-        :param account: account to get recommendations for
-        :param result_count: how many results are suppose to be returned
-        :param sort: possible values: 'rank', 'added', 'released', 'title'
-        """
-        from sickgear import TRAKT_ACCOUNTS
-        if not account or account not in TRAKT_ACCOUNTS or not TRAKT_ACCOUNTS[account].active:
-            raise TraktAuthException('Account missing or disabled')
-        sort = ('rank', sort)[sort in ('rank', 'added', 'released', 'title')]
-        return self._get_show_lists('users/%s/watchlist/shows/%s?extended=full&page=%d&limit=%d' %
-                                    (TRAKT_ACCOUNTS[account].slug, sort, 1, result_count), account=account)
+        return self._get_show_lists(f'shows/recommended/{use_period}?extended=full&page={1:d}&limit={result_count:d}')
 
     def get_anticipated(self, result_count=100, **kwargs):
         # type: (...) -> List[TVInfoShow]
@@ -600,7 +494,7 @@ class TraktIndexer(TVInfoBase):
         get most anticipated shows
         :param result_count: how many results are suppose to be returned
         """
-        return self._get_show_lists('shows/anticipated?extended=full&page=%d&limit=%d' % (1, result_count))
+        return self._get_show_lists(f'shows/anticipated?extended=full&page={1:d}&limit={result_count:d}')
 
     def get_trending(self, result_count=100, **kwargs):
         # type: (...) -> List[TVInfoShow]
@@ -608,7 +502,7 @@ class TraktIndexer(TVInfoBase):
         get trending shows
         :param result_count: how many results are suppose to be returned
         """
-        return self._get_show_lists('shows/trending?extended=full&page=%d&limit=%d' % (1, result_count))
+        return self._get_show_lists(f'shows/trending?extended=full&page={1:d}&limit={result_count:d}')
 
     def get_popular(self, result_count=100, **kwargs):
         # type: (...) -> List[TVInfoShow]
@@ -616,7 +510,7 @@ class TraktIndexer(TVInfoBase):
         get all popular shows
         :param result_count: how many results are suppose to be returned
         """
-        return self._get_show_lists('shows/popular?extended=full&page=%d&limit=%d' % (1, result_count))
+        return self._get_show_lists(f'shows/popular?extended=full&page={1:d}&limit={result_count:d}')
 
     def get_similar(self, tvid, result_count=100, **kwargs):
         # type: (integer_types, int, Any) -> List[TVInfoShow]
@@ -627,7 +521,7 @@ class TraktIndexer(TVInfoBase):
         """
         if not isinstance(tvid, int):
             raise TraktException('tvid/trakt id for show required')
-        return self._get_show_lists('shows/%d/related?extended=full&page=%d&limit=%d' % (tvid, 1, result_count))
+        return self._get_show_lists(f'shows/{tvid:d}/related?extended=full&page={1:d}&limit={result_count:d}')
 
     def get_new_shows(self, result_count=100, start_date=None, days=32, **kwargs):
         # type: (...) -> List[TVInfoShow]
@@ -639,8 +533,8 @@ class TraktIndexer(TVInfoBase):
         """
         if None is start_date:
             start_date = (datetime.datetime.now() + datetime.timedelta(days=-16)).strftime('%Y-%m-%d')
-        return self._get_show_lists('calendars/all/shows/new/%s/%s?extended=full&page=%d&limit=%d' %
-                                    (start_date, days, 1, result_count))
+        return self._get_show_lists(
+            f'calendars/all/shows/new/{start_date}/{days}?extended=full&page={1:d}&limit={result_count:d}')
 
     def get_new_seasons(self, result_count=100, start_date=None, days=32, **kwargs):
         # type: (...) -> List[TVInfoShow]
@@ -652,5 +546,5 @@ class TraktIndexer(TVInfoBase):
         """
         if None is start_date:
             start_date = (datetime.datetime.now() + datetime.timedelta(days=-16)).strftime('%Y-%m-%d')
-        return self._get_show_lists('calendars/all/shows/premieres/%s/%s?extended=full&page=%d&limit=%d' %
-                                    (start_date, days, 1, result_count))
+        return self._get_show_lists(
+            f'calendars/all/shows/premieres/{start_date}/{days}?extended=full&page={1:d}&limit={result_count:d}')