+++ sickgear/webserve.py
@@ -71,7 +71,6 @@ from .sgdatetime import SGDatetime
 from .show_name_helpers import abbr_showname
 
 from .show_updater import clean_ignore_require_words
-from .trakt_helpers import build_config, trakt_collection_remove_account
 from .tv import TVidProdid, Person as TVPerson, Character as TVCharacter, TVSWITCH_NORMAL, tvswitch_names, \
     TVSWITCH_EP_DELETED, tvswitch_ep_names, usable_id
 
@@ -86,7 +85,7 @@ from tornado.escape import utf8
 from tornado.web import RequestHandler, StaticFileHandler, authenticated
 from tornado.concurrent import run_on_executor
 
-from lib import requests
+# from lib import requests
 from lib.urllib3.util.retry import Retry
 from lib.requests.adapters import HTTPAdapter
 from lib.iso639 import Lang
@@ -99,7 +98,6 @@ try:
     from lib.thefuzz import fuzz
 except ImportError as e:
     from lib.fuzzywuzzy import fuzz
-from lib.api_trakt import TraktAPI
 from lib.api_trakt.exceptions import TraktException, TraktAuthException
 from lib.api_imdb.imdb_exceptions import IMDbException
 from lib.tvinfo_base import TVInfoEpisode, RoleTypes
@@ -142,7 +140,7 @@ else:
 
 # noinspection PyUnreachableCode
 if False:
-    from typing import Any, AnyStr, Dict, List, Optional, Set, Tuple, Union
+    from typing import Any, AnyStr, Callable, Dict, List, Optional, Set, Tuple, Union
     from sickgear.providers.generic import TorrentProvider
     # prevent pyc TVInfoBase resolution by typing the derived used class to TVInfoAPI instantiation
     from lib.tvinfo_base import TVInfoBase, TVInfoCharacter, TVInfoPerson
@@ -2077,55 +2075,6 @@ class Home(MainHandler):
 
         return notifiers.NotifierFactory().get('LIBNOTIFY').test_notify()
 
-    def trakt_authenticate(self, pin=None, account=None):
-        self.set_header('Cache-Control', 'max-age=0,no-cache,no-store')
-
-        if None is pin:
-            return json_dumps({'result': 'Fail', 'error_message': 'Trakt PIN required for authentication'})
-
-        if account and 'new' == account:
-            account = None
-
-        acc = None
-        if account:
-            acc = helpers.try_int(account, -1)
-            if 0 < acc and acc not in sickgear.TRAKT_ACCOUNTS:
-                return json_dumps({'result': 'Fail', 'error_message': 'Fail: cannot update non-existing account'})
-
-        json_fail_auth = json_dumps({'result': 'Fail', 'error_message': 'Trakt NOT authenticated'})
-        try:
-            resp = TraktAPI().trakt_token(pin, account=acc)
-        except TraktAuthException:
-            return json_fail_auth
-        if not account and isinstance(resp, bool) and not resp:
-            return json_fail_auth
-
-        if not sickgear.USE_TRAKT:
-            sickgear.USE_TRAKT = True
-            sickgear.save_config()
-        pick = resp if not account else acc
-        return json_dumps({'result': 'Success',
-                           'account_id': sickgear.TRAKT_ACCOUNTS[pick].account_id,
-                           'account_name': sickgear.TRAKT_ACCOUNTS[pick].name})
-
-    def trakt_delete(self, accountid=None):
-        self.set_header('Cache-Control', 'max-age=0,no-cache,no-store')
-
-        if accountid:
-            aid = helpers.try_int(accountid, None)
-            if None is not aid:
-                if aid in sickgear.TRAKT_ACCOUNTS:
-                    account = {'result': 'Success',
-                               'account_id': sickgear.TRAKT_ACCOUNTS[aid].account_id,
-                               'account_name': sickgear.TRAKT_ACCOUNTS[aid].name}
-                    if TraktAPI.delete_account(aid):
-                        trakt_collection_remove_account(aid)
-                        account['num_accounts'] = len(sickgear.TRAKT_ACCOUNTS)
-                        return json_dumps(account)
-
-                return json_dumps({'result': 'Not found: Account to delete'})
-        return json_dumps({'result': 'Not found: Invalid account id'})
-
     def load_show_notify_lists(self):
         self.set_header('Cache-Control', 'max-age=0,no-cache,no-store')
 
@@ -3277,10 +3226,6 @@ class Home(MainHandler):
         if sickgear.show_queue_scheduler.action.is_being_added(
                 show_obj) or sickgear.show_queue_scheduler.action.is_being_updated(show_obj):
             return self._generic_message("Error", "Shows can't be deleted while they're being added or updated.")
-
-        # if sickgear.USE_TRAKT and sickgear.TRAKT_SYNC:
-        #     # remove show from trakt.tv library
-        #     sickgear.trakt_checker_scheduler.action.removeShowFromTraktLibrary(show_obj)
 
         show_obj.delete_show(bool(full))
 
@@ -5667,7 +5612,7 @@ class AddShows(Home):
     @private_call
     @staticmethod
     def allow_browse_mru(mode_or_mru):
-        # Fix an issue where a default view mixed with a deriviative view that requires a param will break the default
+        # Fix an issue where a default view mixed with a derivative view that requires a param will break the default
         # Disallows default views from using derivative mru's
         return 'person' not in mode_or_mru
 
@@ -5832,10 +5777,18 @@ class AddShows(Home):
             return self.new_show('|'.join(['', '', '', ' '.join([ids, show_name])]), use_show_name=True)
 
     def trakt_default(self):
-        method = getattr(self, sickgear.TRAKT_MRU, None)
-        if not callable(method) or not self.allow_browse_mru(sickgear.TMDB_MRU):
-            return self.trakt_trending()
-        return method()
+        if sickgear.TRAKT_MRU:
+            try:
+                method, args = sickgear.TRAKT_MRU.split('?')
+                kwargs = dict(args.split('=') for _ in args.strip().splitlines())
+            except (BaseException, Exception):
+                method, kwargs = sickgear.TRAKT_MRU, {}
+
+            func = getattr(self, method, None)  # type: Callable
+            if func and callable(func) and self.allow_browse_mru(method):
+                return func(**kwargs)
+
+        return self.trakt_trending()
 
     def trakt_anticipated(self):
 
@@ -5897,36 +5850,6 @@ class AddShows(Home):
             f'Most {action}ed at Trakt during the last {desc}',
             mode=f'{action}ed{ext}', period=f'{cycle}ly')
 
-    def trakt_recommended(self, **kwargs):
-
-        if 'add' == kwargs.get('action'):
-            return self.redirect('/config/notifications/#tabs-3')
-
-        account = helpers.try_int(kwargs.get('account'))
-        try:
-            name = sickgear.TRAKT_ACCOUNTS[account].name
-        except KeyError:
-            return self.trakt_default()
-        return self.browse_trakt(
-            'get_recommended_for_account',
-            'Recommended for <b class="grey-text">%s</b> by Trakt' % name,
-            mode='recommended-%s' % account, account=account, ignore_collected=True, ignore_watchlisted=True)
-
-    def trakt_watchlist(self, **kwargs):
-
-        if 'add' == kwargs.get('action'):
-            return self.redirect('/config/notifications/#tabs-3')
-
-        account = helpers.try_int(kwargs.get('account'))
-        try:
-            name = sickgear.TRAKT_ACCOUNTS[account].name
-        except KeyError:
-            return self.trakt_default()
-        return self.browse_trakt(
-            'get_watchlisted_for_account',
-            'WatchList for <b class="grey-text">%s</b> by Trakt' % name,
-            mode='watchlist-%s' % account, account=account, ignore_collected=True)
-
     @private_call
     def get_trakt_data(self, api_method, **kwargs):
 
@@ -5959,13 +5882,6 @@ class AddShows(Home):
                 items = t.get_most_played(**kwargs)
             elif 'get_most_collected' == api_method:
                 items = t.get_most_collected(**kwargs)
-            elif 'get_recommended_for_account' == api_method:
-                items = t.get_recommended_for_account(**kwargs)
-            elif 'get_watchlisted_for_account' == api_method:
-                items = t.get_watchlisted_for_account(**kwargs)
-                if not items:
-                    error_msg = 'No items in watchlist.  Use the "Add to watchlist" button at the Trakt website'
-                    raise ValueError(error_msg)
             elif 'person' == mode:
                 items = []
                 p_item = t.get_person(get_show_credits=True, **kwargs)  # type: TVInfoPerson
@@ -6097,11 +6013,11 @@ class AddShows(Home):
         browse_type = 'Trakt'
         mode = kwargs.get('mode', '')
         filtered = []
-        all_classes = set()
 
         if not sickgear.USE_TRAKT \
                 and ('recommended' in mode or 'watchlist' in mode):
-            error_msg = 'To browse personal recommendations, enable Trakt.tv in Config/Notifications/Social'
+            # error_msg = 'To browse personal recommendations, enable Trakt.tv in Config/Notifications/Social'
+            error_msg = 'Deprecated due to VIP only; personal recommendations'
             return self.browse_shows(browse_type, browse_title, filtered, error_msg=error_msg, show_header=1, **kwargs)
 
         try:
@@ -6116,7 +6032,8 @@ class AddShows(Home):
 
         kwargs.update(dict(oldest=oldest, newest=newest, error_msg=error_msg, use_networks=use_networks))
 
-        if not any(m in mode for m in ('recommended', 'watchlist', 'person')):
+        # if not any(m in mode for m in ('recommended', 'watchlist', 'person')):
+        if not any(m in mode for m in ('person',)):
             mode = mode.split('-')
             if mode and self.allow_browse_mru(mode):
                 func = 'trakt_%s' % mode[0]
@@ -10165,10 +10082,6 @@ class ConfigNotifications(Config):
             use_libnotify=None, libnotify_notify_onsnatch=None, libnotify_notify_ondownload=None,
             libnotify_notify_onsubtitledownload=None,
 
-            use_trakt=None,
-            # trakt_pin=None, trakt_remove_watchlist=None, trakt_use_watchlist=None, trakt_method_add=None,
-            # trakt_start_paused=None, trakt_sync=None, trakt_default_indexer=None, trakt_remove_serieslist=None,
-            # trakt_collection=None, trakt_accounts=None,
             use_slack=None, slack_notify_onsnatch=None, slack_notify_ondownload=None,
             slack_notify_onsubtitledownload=None, slack_access_token=None, slack_channel=None,
             slack_as_authed=None, slack_bot_name=None, slack_icon_url=None,
@@ -10288,17 +10201,6 @@ class ConfigNotifications(Config):
         sickgear.SYNOLOGYNOTIFIER_NOTIFY_ONDOWNLOAD = config.checkbox_to_value(synologynotifier_notify_ondownload)
         sickgear.SYNOLOGYNOTIFIER_NOTIFY_ONSUBTITLEDOWNLOAD = config.checkbox_to_value(
             synologynotifier_notify_onsubtitledownload)
-
-        sickgear.USE_TRAKT = config.checkbox_to_value(use_trakt)
-        sickgear.TRAKT_UPDATE_COLLECTION = build_config(**kwargs)
-        # sickgear.trakt_checker_scheduler.silent = not sickgear.USE_TRAKT
-        # sickgear.TRAKT_DEFAULT_INDEXER = int(trakt_default_indexer)
-        # sickgear.TRAKT_SYNC = config.checkbox_to_value(trakt_sync)
-        # sickgear.TRAKT_USE_WATCHLIST = config.checkbox_to_value(trakt_use_watchlist)
-        # sickgear.TRAKT_METHOD_ADD = int(trakt_method_add)
-        # sickgear.TRAKT_REMOVE_WATCHLIST = config.checkbox_to_value(trakt_remove_watchlist)
-        # sickgear.TRAKT_REMOVE_SERIESLIST = config.checkbox_to_value(trakt_remove_serieslist)
-        # sickgear.TRAKT_START_PAUSED = config.checkbox_to_value(trakt_start_paused)
 
         sickgear.USE_SLACK = config.checkbox_to_value(use_slack)
         sickgear.SLACK_NOTIFY_ONSNATCH = config.checkbox_to_value(slack_notify_onsnatch)